Sweepstakes class has been further modified to include RegisterContestant, PickWinner, and PrintContestantInfo methods.

+++ SweepStakesProject/Contestant.cs
@@ -11,7 +11,7 @@ namespace SweepStakesProject
         public string FirstName; 
         public string LastName;
         public string EmailAddress;
-        public int RegistrationNumber;
+        public double RegistrationNumber;
         public Contestant()
         {
             FirstName = "";
